+++ datafusion/expr/src/lib.rs
@@ -71,7 +71,7 @@ pub use function::{
 pub use datafusion_expr_common::groups_accumulator::{EmitTo, GroupsAccumulator};
 pub use literal::{lit, lit_timestamp_nano, Literal, TimestampLiteral};
 pub use logical_plan::*;
-pub use operator::Operator;
+pub use datafusion_expr_common::operator::Operator;
 pub use partition_evaluator::PartitionEvaluator;
 pub use datafusion_expr_common::signature::{
     ArrayFunctionSignature, FuncMonotonicity, Signature, TypeSignature, Volatility,
